+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changelog
 * Fixed function name overlapping information icon
 * Clear all data will also clear local caches
 * Show parameters in usage data
+* Added restriction for deprecated *contacts/people*
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)